Теперь байт пикселя хранит вариацию цвета

+++ src/MainApp.java
@@ -19,9 +19,9 @@ public class MainApp {
             MainLoop mainLoop = new MainLoop(window, panel, mainRenderer);
         });
 
-        mainGrid.setPixel(100, 100, (byte) 1);
-        mainGrid.setPixel(120, 110, (byte) 1);
-        mainGrid.setPixel(150, 80, (byte) 1);
+        mainGrid.setPixel(100, 100, 1);
+        mainGrid.setPixel(120, 110, 1);
+        mainGrid.setPixel(150, 80, 1);
 
     }
 }
